requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-SocketIO==5.3.6
Flask-Login==0.6.3
Werkzeug==2.3.7
python-socketio==5.8.0
python-engineio==4.7.1
python-dotenv==1.0.0
